Update tests for new WebAssembly.Global API
The current version of the WebAssembly JS API specification [1]
exports globals as WebAssembly.Global, rather than as Numbers. This
version of the API is shipped in Chrome 69 and Firefox 62, and is
also seen in Node 11 nightly builds.

This patch updates webpack's tests to permit either the new API or
the old one, for exports of WebAssembly Globals. One of the affected
tests passes on Node 11 (where it previously failed), and the other
fails for unrelated reasons [2].

Fixes #8145

[1] https://webassembly.github.io/spec/js-api/index.html
[2] https://github.com/webpack/webpack/pull/7467/files#r223454525